Added platform listing and viewing
